bsps/shared/dev: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This is frequently an int or ssize_t variable compared to a uint32_t or
size_t. Sometimes the size_t is from a sizeof() use.

+++ bsps/shared/dev/flash/flash_sim_flashdev.c
@@ -160,7 +160,7 @@ static int flash_sim_erase_wrapper(
   uintptr_t aligned_end = RTEMS_ALIGN_UP( offset + count, sector_size );
   uint64_t sector_count = ( aligned_end - aligned_start ) / sector_size;
 
-  for ( int i = 0; i < sector_count; i++ ) {
+  for ( uint64_t i = 0; i < sector_count; i++ ) {
     erase_sector( flash_driver, aligned_start + i * sector_size );
   }
